Fix imports in tests
