+++ manticore/native/__init__.py
@@ -1,3 +1,6 @@
 from manticore.utils.install_helper import ensure_native_deps
 
 ensure_native_deps()
+
+from .manticore import Manticore
+from ..utils.helpers import issymbolic
