+++ manticore/native/manticore.py
@@ -75,6 +75,20 @@ class Manticore(ManticoreBase):
         except KeyError:  # FIXME(mark) magic parsing for DECREE should raise better error
             raise Exception(f'Invalid binary: {path}')
 
+    def invoke_model(self, model):
+        '''
+        Invoke a `model`. A `model` is a callable whose first argument is a
+        :class:`~manticore.core.state.State`. If the `model` models a normal (non-variadic)
+        function, the following arguments correspond to the arguments of the C function
+        being modeled. If the `model` models a variadic function, the following argument
+        is a generator object, which can be used to access function arguments dynamically.
+        The `model` callable should simply return the value that should be returned by the
+        native function being modeled.
+
+        :param callable model: Model to invoke
+        '''
+        self._platform.invoke_model(model, prefix_args=(self,))
+
     #############################################################################
     #############################################################################
     #############################################################################
