+++ tests/eth_detectors.py
@@ -7,14 +7,14 @@ import unittest
 
 import os
 import shutil
-from eth_general import make_mock_evm_state
 
 from manticore.core.smtlib import operators
 from manticore.ethereum import DetectIntegerOverflow, DetectUnusedRetVal, DetectSelfdestruct, \
     DetectDelegatecall, \
     DetectExternalCallAndLeak, DetectEnvInstruction, DetectRaceCondition
-from ethereum.manticore import ManticoreEVM
-from ethereum.plugins import LoopDepthLimiter
+from manticore.ethereum.manticore import ManticoreEVM
+from manticore.ethereum.plugins import LoopDepthLimiter
+from .eth_general import make_mock_evm_state
 
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
