+++ tests/test_dyn.py
@@ -2,7 +2,7 @@
 import unittest
 from manticore.core.cpu.x86 import *
 from manticore.core.smtlib import Operators
-from native.memory import *
+from manticore.native.memory import *
 
 
 class CPUTest(unittest.TestCase):
