+++ tests/test_rlp.py
@@ -5,9 +5,11 @@ import doctest
 
 from manticore.utils import rlp
 
+
 def load_tests(loader, tests, ignore):
     tests.addTests(doctest.DocTestSuite(rlp))
     return tests    
+
 
 if __name__ == '__main__':
     import unittest
